@@ -11,19 +11,24 @@
     "build": "node build/build.js"
   },
   "dependencies": {
-    "axios": "^0.18.0",
-    "bootstrap": "^4.1.3",
-    "bootstrap-vue": "^2.0.0-rc.11",
+    "ajv": "^6.10.2",
+    "axios": "^0.18.1",
+    "bootstrap": "^4.3.1",
+    "bootstrap-vue": "^2.0.4",
     "ejs": "^2.6.1",
     "express": "^4.16.3",
     "i": "^0.3.6",
     "iview": "^3.1.3",
+    "jquery": "^3.4.1",
+    "js-yaml": "^3.13.1",
+    "popper.js": "^1.16.0",
+    "portal-vue": "^2.1.6",
     "pure": "^2.85.0",
     "pure-vue": "0.0.0",
     "socket.io": "^2.1.1",
     "socket.io-client": "^2.1.1",
-    "video.js": "7.0.5",
-    "vue": "^2.5.2",
+    "video.js": "^7.6.5",
+    "vue": "^2.6.10",
     "vue-chat-scroll": "^1.3.3",
     "vue-moment": "^4.0.0",
     "vue-video-player": "^5.0.2"
@@ -40,13 +45,13 @@
     "babel-preset-stage-2": "^6.22.0",
     "chalk": "^2.0.1",
     "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.0",
+    "css-loader": "^3.2.0",
     "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^1.1.4",
     "friendly-errors-webpack-plugin": "^1.6.1",
     "html-webpack-plugin": "^2.30.1",
     "node-notifier": "^5.1.2",
-    "optimize-css-assets-webpack-plugin": "^3.2.0",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
     "ora": "^1.2.0",
     "portfinder": "^1.0.13",
     "postcss-import": "^11.0.0",
@@ -59,10 +64,10 @@
     "url-loader": "^1.1.2",
     "vue-loader": "^13.3.0",
     "vue-style-loader": "^3.0.1",
-    "vue-template-compiler": "^2.5.2",
+    "vue-template-compiler": "^2.6.10",
     "webpack": "^3.6.0",
-    "webpack-bundle-analyzer": "^2.9.0",
-    "webpack-dev-server": "^2.9.1",
+    "webpack-bundle-analyzer": "^3.6.0",
+    "webpack-dev-server": "^2.11.5",
     "webpack-merge": "^4.1.0"
   },
   "engines": {
