@@ -1,10 +1,6 @@
 <template>
-    <div>  
-        <div class="videocontent">
-          <video  
-              id="player"
-                class="video-js vjs-default-skin vjs-big-play-centered">
-              <p class="vjs-no-js">
+          <video ref="videoPlayer" class="video-js">
+                <p class="vjs-no-js">
                 To view this video please enable JavaScript, and consider upgrading to a
                 web browser that
                 <a href="http://videojs.com/html5-video-support/" target="_blank">
@@ -12,8 +8,6 @@
                 </a>
               </p>
           </video>
-        </div>
-    </div>
 </template>
 
 <script>
@@ -21,24 +15,58 @@
   //import videojs from 'video.js'
   import videojstyle from 'video.js/dist/video-js.css' // tuodaan videojs tyylitiedosto
 
-  window.videojs = require('video.js');
+  //window.videojs = require('video.js');
+  import videojs from 'video.js'
 
   export default {
   name: 'Videoplayer',
-  props: ['videos', 'newUrl'],
-  data() {
-    return {
-      player: '',
-      defaultSrc: 'http://wowza.oubs.fi/vod/mp4:sample.mp4/playlist.m3u8',
-      playerOptions: {
-        autoplay: false,
-        controls: true,
-        muted: true,
-        fluid: true,
-        aspectRatio: "16:9",
-        poster: "https://oubs.fi/wp-content/uploads/2016/11/cropped-oubs_white_teksti_lapi_152px.png"
+  props: {
+    newUrl: {
+      type: String,
+      default() {
+        return {
+          newUrl: newUrl
+        }
+      }
+    },
+    videos: {
+      type: Array,
+      default() {
+        return {
+          videos: videos
+        }
+      }
+    },
+    options: {
+      type: Object,
+      default() {
+        return {
+            autoplay: false,
+            controls: true,
+            muted: true,
+            fluid: true,
+            aspectRatio: "16:9",
+            poster: "https://oubs.fi/wp-content/uploads/2016/11/cropped-oubs_white_teksti_lapi_152px.png"
+        }
       }
     }
+  },
+  
+  data() {
+    return {
+      player: null,
+    }
+  },
+  mounted() {
+        this.player = videojs(this.$refs.videoPlayer, this.options, function onPlayerReady() {
+            console.log('onPlayerReady', this);
+        })
+        this.player.src('http://wowza.oubs.fi/vod/mp4:sample.mp4/playlist.m3u8');
+    },
+  beforeDestroy() {
+      if (this.player) {
+            this.player.dispose()
+      }
   },
 
   watch: {
@@ -57,16 +85,6 @@
         this.player.dispose(); 
     },
   },
-
-  mounted() {
-    //window.playerEvents = this;
-    this.playerInit();
-    this.player.src(this.defaultSrc);
-  },
-
-  beforeDestroy() { 
-        this.playerDispose(); 
-    }
 };
 
 </script>
